requirements.txt
Flask
Flask_SQLAlchemy
gunicorn
psycopg2-binary
Werkzeug
